JDBC-33: Add optional connectTimeout
If the connectTimeout property is set, after the Cluster is created
the JDBC driver will perform the "waitUntilReady" operation with that
given timeout, making sure that it is eagerly checked if the connections
can be established.

If the connectTimeout is not provided, no waitUntilReady is performed
and then as  result the query will time out at the timeout interval,
likely 75s.