JDBCC-44: Tableau requires restart if SSL parameters are changed
Details:
Currently only one cluster environment is created which is shared
by all the connections. Therefore, even if a new connection request
arrives with only ssl parameters changed, the driver uses the
existing cluster environment contaning the security config that was
created for the first connection. With this change we create separate
cluster environment for separate connections.

Change-Id: Ie203a3caa48912eb69999663b0fd39c5059090b1
Reviewed-on: https://review.couchbase.org/c/couchbase-jdbc-driver/+/185978
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Peeyush Gupta <peeyush.gupta@couchbase.com>
Reviewed-by: Ali Alsuliman <ali.alsuliman@couchbase.com>