JDBCC-43: Tableau query dialog can't be closed
Tableau query dailogue hangs if the Couchbase server is down.
The reason being that Couchbase JDBC driver sets infinite query
execution timeout by default (as mentioned in JDBC spec). However,
for prepareStatement API that sends query to analytics with compileOnly
set to true, there is no need for infinite timeout, and the default
75s timeout set by the Java SDK works.

Change-Id: I91981814a8a6ebdbaf7569bb2551d8d4685b9659
Reviewed-on: https://review.couchbase.org/c/couchbase-jdbc-driver/+/188681
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Peeyush Gupta <peeyush.gupta@couchbase.com>
Reviewed-by: Peeyush Gupta <peeyush.gupta@couchbase.com>
Reviewed-by: Ali Alsuliman <ali.alsuliman@couchbase.com>